Plot Break Down Objects in D3 with r2d3 package.

Plot Break Down Objects in D3 with r2d3 package.

plotD3(x, ...)

# S3 method for break_down
plotD3(x, ..., baseline = NA, max_features = 10,
  min_max = NA, vcolors = DALEX::theme_drwhy_colors_break_down(),
  digits = 3, rounding_function = round)

Arguments

x
the model model of `break_down`` class.

...
other parameters.

baseline
if numeric then veritical line will start in baseline.

max_features
maximal number of features to be included in the plot. default value is 10.

min_max
a range of OX axis. By deafult `NA` therefore will be extracted from the contributions of `x`. But can be set to some constants, usefull if these plots are used for comparisons.

vcolors
named vector with colors.

digits
number of decimal places (round) or significant digits (signif) to be used. See the rounding_function argument.

rounding_function
function that is to used for rounding numbers. It may be signif which keeps a specified number of significant digits. Or the default round to have the same precision for all components.

Value

an `r2d3` object.

Examples

## Not run:
# prepare dataset
library("titanic")
titanic <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                            "SibSp", "Parch", "Fare", "Embarked")]
titanic$Survived <- factor(titanic$Survived)
titanic$Sex <- factor(titanic$Sex)
titanic$Embarked <- factor(titanic$Embarked)
titanic <- na.omit(titanic)

# prepare model
library("randomForest")
rf_model <- randomForest(Survived ~ .,  data = titanic)

# prepare explainer
library("DALEX")
rf_explain <- explain(rf_model, data = titanic,
                      y = titanic$Survived == "1", label = "RF")

# plor D3 explainers
library("iBreakDown")
rf_la <- local_attributions(rf_explain, titanic[2,])
rf_la

#>                  contribution
#> RF: intercept           0.364
#> RF: Sex = female        0.296
#> RF: Fare = 71           0.069
#> RF: Pclass = 1          0.223
#> RF: Embarked = C        0.010
#> RF: Age = 38            0.014
#> RF: SibSp = 1           0.013
#> RF: Parch = 0           0.007
#> RF: Survived = 1        0.000
#> RF: prediction          0.996

plotD3(rf_la)

rf_la <- local_attributions(rf_explain, titanic[3,])
rf_la

#>                  contribution
#> RF: intercept           0.364
#> RF: Sex = female        0.296
#> RF: Pclass = 3         -0.218
#> RF: Fare = 7.9         -0.064
#> RF: SibSp = 0           0.028
#> RF: Embarked = S       -0.029
#> RF: Age = 26            0.123
#> RF: Parch = 0          -0.030
#> RF: Survived = 1        0.000
#> RF: prediction          0.470

plotD3(rf_la, max_features = 4)
plotD3(rf_la, max_features = 4, min_max = c(0,1))