Plot Shap (Break Down Uncertainty) Objects in D3 with r2d3 package.

Plots shapley values. Usually executed after `shap()` function.

# S3 method for shap
plotD3(x, ..., baseline = NA, max_features = 10,
  digits = 3, rounding_function = round, bar_width = 12,
  margin = 0.2, scale_height = FALSE, min_max = NA, vcolors = NA,
  chartTitle = NA, time = 0)

Arguments

x
the model model of `shap` class.

...
other parameters.

baseline
if numeric then veritical line will start in baseline.

max_features
maximal number of features to be included in the plot. Default is 10.

digits
number of decimal places (round) or significant digits (signif) to be used. See the rounding_function argument.

rounding_function
function that is to used for rounding numbers. It may be signif which keeps a specified number of significant digits. Or the default round to have the same precision for all components.

bar_width
width of bars in px. Default is 12px

margin
extend x axis domain range to adjust the plot. Usually value between 0.1 and 0.3, by default it's 0.2

scale_height
should the height of plot scale with window size? By default it's FALSE

min_max
a range of OX axis. By deafult `NA` therefore will be extracted from the contributions of `x`. But can be set to some constants, usefull if these plots are used for comparisons.

vcolors
named vector with colors. By default `NA` therfore will choose DrWhy colors

chartTitle
a character. Set custom title

time
in ms. Set animation length

Value

an `r2d3` object.

References

Predictive Models: Visual Exploration, Explanation and Debugging https://pbiecek.github.io/PM_VEE

Examples

library("DALEX")
library("iBreakDown")

titanic <- na.omit(titanic)
set.seed(1313)
titanic_small <- titanic[sample(1:nrow(titanic), 500),]
model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class +
                           embarked + country + sibsp + parch,
                         data = titanic_small, family = "binomial")
explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic_small[,-9],
                               y = titanic_small$survived == "yes",
                               label = "glm")

#> Preparation of a new explainer is initiated
#>   -> model label       :  glm 
#>   -> data              :  500  rows  8  cols 
#>   -> target variable   :  500  values 
#>   -> predict function  :  yhat.glm  will be used ([33mdefault[39m)
#>   -> predicted values  :  numerical, min =  4.748259e-09 , mean =  0.308 , max =  1  
#>   -> residual function :  difference between y and yhat ([33mdefault[39m)
#>   -> residuals         :  numerical, min =  -0.9480574 , mean =  -1.1981e-09 , max =  0.9576496  
#> [32mA new explainer has been created![39m

s_glm <- shap(explain_titanic_glm, titanic_small[1, ])
s_glm

#>                                         min           q1       median
#> glm: age = 36                 -0.0330420958 -0.023847845 -0.023801953
#> glm: class = victualling crew  0.0379950017  0.046376537  0.062364754
#> glm: country = England        -0.0026614617  0.001730294  0.048961160
#> glm: embarked = Southampton   -0.0199796377 -0.015116715 -0.011619284
#> glm: fare = 0                  0.0009373315  0.001319203  0.001359069
#> glm: gender = female           0.4197542885  0.433079924  0.464561634
#> glm: parch = 0                -0.0097543030 -0.008829139 -0.007969768
#> glm: sibsp = 0                -0.0015993005 -0.001299609 -0.001190244
#>                                       mean           q3           max
#> glm: age = 36                 -0.021927670 -0.013335539 -0.0115739621
#> glm: class = victualling crew  0.067971650  0.091421908  0.1018895144
#> glm: country = England         0.034546339  0.060992270  0.0617575237
#> glm: embarked = Southampton   -0.013072235 -0.011166947 -0.0106583875
#> glm: fare = 0                  0.001339487  0.001402580  0.0014461837
#> glm: gender = female           0.464262110  0.475641674  0.5237532113
#> glm: parch = 0                -0.007926482 -0.007131963 -0.0059009128
#> glm: sibsp = 0                -0.001204145 -0.001075244 -0.0009629568

plotD3(s_glm)

library(randomForest)

HR_small <- HR[2:500,]
m_rf <- randomForest(status ~. , data = HR_small)
new_observation <- HR_test[1,]
new_observation

#>   gender      age    hours evaluation salary status
#> 1   male 57.72683 42.31527          2      2  fired

p_fun <- function(object, newdata){predict(object, newdata=newdata, type = "prob")}

s_rf <- shap(m_rf,
             data = HR_small[,-6],
             new_observation =  new_observation,
             predict_function = p_fun)

plotD3(s_rf, time = 500)