Plot Shap (Break Down Uncertainty) Objects in D3 with r2d3 package.

Plots shapley values.

# S3 method for shap
plotD3(x, ..., baseline = NA, max_features = 10,
  digits = 3, rounding_function = round, bar_width = 12,
  margin = 0.2, scale_height = FALSE, min_max = NA, vcolors = NA,
  chart_title = NA, time = 0)

Arguments

x
an explanation created with shap

...
other parameters.

baseline
if numeric then veritical line will start in baseline.

max_features
maximal number of features to be included in the plot. Default is 10.

digits
number of decimal places (round) or significant digits (signif) to be used. See the rounding_function argument.

rounding_function
a function to be used for rounding numbers. This should be signif which keeps a specified number of significant digits or round (which is default) to have the same precision for all components.

bar_width
width of bars in px. Default is 12px

margin
extend x axis domain range to adjust the plot. Usually value between 0.1 and 0.3, by default it's 0.2

scale_height
if TRUE, the height of the plot scales with window size.

min_max
a range of OX axis. By deafult `NA` therefore will be extracted from the contributions of `x`. But can be set to some constants, usefull if these plots are used for comparisons.

vcolors
If NA (default), DrWhy colors are used.

chart_title
a character. Set custom title

time
in ms. Set animation length

Value

a `r2d3` object.

References

Predictive Models: Visual Exploration, Explanation and Debugging https://pbiecek.github.io/PM_VEE

Examples

library("DALEX")
library("iBreakDown")

titanic <- na.omit(titanic)
set.seed(1313)
titanic_small <- titanic[sample(1:nrow(titanic), 500),]
model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class +
                           embarked + country + sibsp + parch,
                         data = titanic_small, family = "binomial")
explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic_small[,-9],
                               y = titanic_small$survived == "yes",
                               label = "glm")

#> Preparation of a new explainer is initiated
#>   -> model label       :  glm 
#>   -> data              :  500  rows  8  cols 
#>   -> target variable   :  500  values 
#>   -> predict function  :  yhat.glm  will be used (default)
#>   -> predicted values  :  numerical, min =  7.022097e-10 , mean =  0.298 , max =  1  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.9336566 , mean =  -1.160845e-09 , max =  0.9752572  
#> A new explainer has been created!

s_glm <- shap(explain_titanic_glm, titanic_small[1, ])
s_glm

#>                                       min            q1       median
#> glm: age = 50               -1.198449e-01 -0.1106659660 -0.095308733
#> glm: class = 2nd            -5.298315e-02 -0.0489164568 -0.043860368
#> glm: country = England      -1.659964e-02 -0.0155494537 -0.009284810
#> glm: embarked = Southampton -1.779786e-02 -0.0134423905 -0.013146172
#> glm: fare = 13               9.257581e-05  0.0005037324  0.001682814
#> glm: gender = male          -1.209775e-01 -0.1094698969 -0.101361558
#> glm: parch = 0              -5.518703e-03 -0.0042501528 -0.003879446
#> glm: sibsp = 0               4.007248e-03  0.0047186704  0.010359019
#>                                     mean           q3           max
#> glm: age = 50               -0.100731328 -0.092665631 -0.0827366764
#> glm: class = 2nd            -0.043962169 -0.039051289 -0.0354949542
#> glm: country = England      -0.008667403 -0.002271641 -0.0017944816
#> glm: embarked = Southampton -0.011409841 -0.007864790 -0.0053594363
#> glm: fare = 13               0.001465891  0.002146564  0.0028319087
#> glm: gender = male          -0.099395950 -0.095731684 -0.0762848374
#> glm: parch = 0              -0.003409139 -0.002030032 -0.0009953965
#> glm: sibsp = 0               0.012242720  0.020295607  0.0239373088

plotD3(s_glm)

library(randomForest)

HR_small <- HR[2:500,]
m_rf <- randomForest(status ~. , data = HR_small)
new_observation <- HR_test[1,]
new_observation

#>   gender      age    hours evaluation salary status
#> 1   male 57.72683 42.31527          2      2  fired


p_fun <- function(object, newdata){predict(object, newdata=newdata, type = "prob")}

s_rf <- shap(m_rf,
             data = HR_small[,-6],
             new_observation =  new_observation,
             predict_function = p_fun)

plotD3(s_rf, time = 500)